feat: enhance interaction for preview
