+++ packages/rath-client/package.json
@@ -5,7 +5,7 @@
   "homepage": "./",
   "prettier": {
     "tabWidth": 4,
-    "printWidth": 120,
+    "printWidth": 150,
     "singleQuote": true
   },
   "dependencies": {
